+++ src/components/appbar.tsx
@@ -11,7 +11,6 @@ export default function ViewAppBar({
   return (
     <div className="p-8 grid grid-cols-2 gap-4 sticky top-0 z-10 items-center bg-white bg-opacity-80 backdrop-blur-3xl backdrop-saturate-150">
       <div className="branding">{branding}</div>
-
       <div className="flex justify-end gap-4">{actions}</div>
     </div>
   );
