+++ src/components/index.js
@@ -1,4 +1,4 @@
-export { default as AppBar } from "./appbar";
+export { default as AppBar } from "./appBar";
 export { default as Scaffold } from "./scaffold";
 export { default as EventCard } from "./eventCard";
 export { default as LoadingList } from "./loadingList";
